+++ src/main/java/utils/Dispatcher.java
@@ -1,5 +1,7 @@
-package controller;
+package utils;
 
+import controller.DispatchInfo;
+import controller.Handler;
 import http.HttpRequest;
 import http.HttpResponse;
 
@@ -7,12 +9,12 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.Arrays;
 
-public class Controllers {
-    public HttpResponse dispatch(HttpRequest request) throws IOException, URISyntaxException {
+public class Dispatcher {
+    public static HttpResponse dispatch(HttpRequest request) throws IOException, URISyntaxException {
         return findMatchingHandlers(request).handleRequest(request);
     }
 
-    Handler findMatchingHandlers(HttpRequest request) {
+    public static Handler findMatchingHandlers(HttpRequest request) {
         return Arrays.stream(DispatchInfo.values())
                 .filter(dispatchInfo -> dispatchInfo.matchWith(request))
                 .findAny()
